kafka-streaming-project/requirements.txt
##############################
# Kafka & Avro
##############################

fastavro==1.8.3

##############################
# Spark + Delta Lake
##############################
pyspark==3.4.1
delta-spark==2.4.0

##############################
# Data Quality
##############################
great_expectations==0.18.12


##############################
# Monitoring
##############################
prometheus-client==0.19.0

##############################
# Utilities
##############################
faker==18.9.0
pyyaml==6.0.1
pydantic==1.10.13
python-dotenv==1.0.0

##############################
# Testing
##############################
pytest==7.4.2
pytest-mock==3.11.1
pytest-cov==4.1.0
